+++ R/db_ops.R
@@ -147,6 +147,7 @@ rq_colnames <- function(db, table_name,
   # first shot- see if it is a db info with function overrriden
   connection <- db
   connection_options <- NULL
+  q_table_name <- quote_identifier(db, table_name)
   if("rquery_db_info" %in% class(db)) {
     f <- db$rq_colnames
     if(!is.null(f)) {
@@ -154,6 +155,7 @@ rq_colnames <- function(db, table_name,
     }
     connection_options <- db$connection_options
     connection <- db$connection
+    q_table_name <- db$quote_table_name(db, table_name, qualifiers = qualifiers)
   }
   if(is.null(connection)) {
     stop("rquery::rq_colnames db$connection was null")
@@ -161,11 +163,11 @@ rq_colnames <- function(db, table_name,
   # DBI::dbListFields fails intermitnently, and sometimes gives wrong results
   # filed as: https://github.com/tidyverse/dplyr/issues/3204
   if(getDBOption(db, "use_DBI_dbListFields", FALSE, connection_options) && requireNamespace("DBI", quietly = TRUE)) {
-    return(DBI::dbListFields(connection, table_name))
+    return(DBI::dbListFields(connection, q_table_name))
   }
-  # below is going to have issues to to R-column name conversion!
+  # below is going to have issues to R-column name conversion!
   q <- paste0("SELECT * FROM ",
-              quote_identifier(db, table_name),
+              q_table_name,
               " LIMIT 1")
   v <- rq_get_query(db, q)
   colnames(v)
@@ -557,6 +559,12 @@ rq_connection_advice <- function(db) {
     opts[[paste(c("rquery", cname, "use_DBI_dbListFields"), collapse = ".")]] <- FALSE
     opts[[paste(c("rquery", cname, "use_DBI_dbRemoveTable"), collapse = ".")]] <- FALSE
     opts[[paste(c("rquery", cname, "use_DBI_dbExistsTable"), collapse = ".")]] <- FALSE # fails on some CREATE AS tables
+  }
+  if(cname =="PqConnection") { # RPostgres::Postgres()
+    # schema issues
+    opts[[paste(c("rquery", cname, "use_DBI_dbListFields"), collapse = ".")]] <- FALSE
+    opts[[paste(c("rquery", cname, "use_DBI_dbRemoveTable"), collapse = ".")]] <- FALSE
+    opts[[paste(c("rquery", cname, "use_DBI_dbExistsTable"), collapse = ".")]] <- FALSE
   }
   opts
 }