+++ NEWS.md
@@ -1,4 +1,8 @@
 
+# rquery 1.3.4 2019/06/04
+
+ * Work on schema qualifications.
+ 
 # rquery 1.3.3 2019/06/01
 
  * Clean up test_set_indicator() error message.